Refactor Wishlist and ShoppingCart services
Refactored the application's Wishlist and ShoppingCart services to use username as the primary identifier in the addToWishlist and removeFromWishlist methods, rather than userId. Furthermore, significantly improved the functionality of the getAllWishlistItemsByUsername method, ensuring it retrieves a list of all wishlist items for a specific username. Also fine-tuned error handling in ShoppingCartController.